+++ app/src/main/java/com/sevtinge/hyperceiler/module/app/SystemUI/Pad/SystemUiV.java
@@ -24,18 +24,8 @@ import com.sevtinge.hyperceiler.module.base.BaseModule;
 import com.sevtinge.hyperceiler.module.hook.systemui.AllowManageAllNotifications;
 import com.sevtinge.hyperceiler.module.hook.systemui.AutoCollapse;
 import com.sevtinge.hyperceiler.module.hook.systemui.AutoSEffSwitchForSystemUi;
-import com.sevtinge.hyperceiler.module.hook.systemui.BrightnessPct;
-import com.sevtinge.hyperceiler.module.hook.systemui.DisableBottomBar;
-import com.sevtinge.hyperceiler.module.hook.systemui.DisableInfinitymodeGesture;
-import com.sevtinge.hyperceiler.module.hook.systemui.DisableMiuiMultiWinSwitch;
 import com.sevtinge.hyperceiler.module.hook.systemui.DisableTransparent;
-import com.sevtinge.hyperceiler.module.hook.systemui.MonetThemeOverlay;
 import com.sevtinge.hyperceiler.module.hook.systemui.NotificationFix;
-import com.sevtinge.hyperceiler.module.hook.systemui.NotificationFreeform;
-import com.sevtinge.hyperceiler.module.hook.systemui.RemoveMiuiMultiWinSwitch;
-import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.OldWeather;
-import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.media.MediaPicture;
-import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.media.SquigglyProgress;
 import com.sevtinge.hyperceiler.module.hook.systemui.StatusBarActions;
 import com.sevtinge.hyperceiler.module.hook.systemui.StickyFloatingWindowsForSystemUI;
 import com.sevtinge.hyperceiler.module.hook.systemui.UiLockApp;
@@ -51,17 +41,13 @@ import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.FiveGTile;
 import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.FixTilesList;
 import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.GmsTile;
 import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.HideDelimiter;
-import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.media.MediaButton;
-import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.media.MediaControlPanelBackgroundMix;
-import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.media.MediaControlPanelTimeViewTextSize;
-import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.media.MediaSeekBar;
-import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.media.MediaSeekBarColor;
 import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.MuteVisibleNotifications;
 import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.NewFlashLight;
 import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.NotificationColor;
 import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.NotificationImportanceHyperOSFix;
 import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.NotificationRowMenu;
 import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.NotificationWeather;
+import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.OldWeather;
 import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.QQSGrid;
 import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.QSColor;
 import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.QSGrid;
@@ -74,6 +60,13 @@ import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.SunlightMode;
 import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.SunlightModeHigh;
 import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.SwitchCCAndNotification;
 import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.TaplusTile;
+import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.media.MediaButton;
+import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.media.MediaControlPanelBackgroundMix;
+import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.media.MediaControlPanelTimeViewTextSize;
+import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.media.MediaPicture;
+import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.media.MediaSeekBar;
+import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.media.MediaSeekBarColor;
+import com.sevtinge.hyperceiler.module.hook.systemui.controlcenter.media.SquigglyProgress;
 import com.sevtinge.hyperceiler.module.hook.systemui.lockscreen.AllowThirdLockScreenUseFace;
 import com.sevtinge.hyperceiler.module.hook.systemui.lockscreen.BlurButton;
 import com.sevtinge.hyperceiler.module.hook.systemui.lockscreen.ChargingCVP;
@@ -85,12 +78,18 @@ import com.sevtinge.hyperceiler.module.hook.systemui.lockscreen.LinkageAnimCusto
 import com.sevtinge.hyperceiler.module.hook.systemui.lockscreen.LockScreenDoubleTapToSleep;
 import com.sevtinge.hyperceiler.module.hook.systemui.lockscreen.NoPassword;
 import com.sevtinge.hyperceiler.module.hook.systemui.lockscreen.RemoveCamera;
-import com.sevtinge.hyperceiler.module.hook.systemui.lockscreen.RemoveSmartScreen;
 import com.sevtinge.hyperceiler.module.hook.systemui.lockscreen.ScramblePIN;
 import com.sevtinge.hyperceiler.module.hook.systemui.navigation.HandleLineCustom;
 import com.sevtinge.hyperceiler.module.hook.systemui.navigation.HideNavigationBar;
 import com.sevtinge.hyperceiler.module.hook.systemui.navigation.NavigationCustom;
 import com.sevtinge.hyperceiler.module.hook.systemui.navigation.RotationButton;
+import com.sevtinge.hyperceiler.module.hook.systemui.other.BrightnessPct;
+import com.sevtinge.hyperceiler.module.hook.systemui.other.DisableBottomBar;
+import com.sevtinge.hyperceiler.module.hook.systemui.other.DisableInfinitymodeGesture;
+import com.sevtinge.hyperceiler.module.hook.systemui.other.DisableMiuiMultiWinSwitch;
+import com.sevtinge.hyperceiler.module.hook.systemui.other.MonetThemeOverlay;
+import com.sevtinge.hyperceiler.module.hook.systemui.other.NotificationFreeform;
+import com.sevtinge.hyperceiler.module.hook.systemui.other.RemoveMiuiMultiWinSwitch;
 import com.sevtinge.hyperceiler.module.hook.systemui.plugin.NewPluginHelperKt;
 import com.sevtinge.hyperceiler.module.hook.systemui.statusbar.BlurEnable;
 import com.sevtinge.hyperceiler.module.hook.systemui.statusbar.DoubleTapToSleep;
@@ -258,7 +257,6 @@ public class SystemUiV extends BaseModule {
         initHook(new ScramblePIN(), mPrefsMap.getBoolean("system_ui_lock_screen_scramble_pin"));
         initHook(ChargingCVP.INSTANCE, mPrefsMap.getBoolean("system_ui_lock_screen_show_charging_cv"));
         initHook(RemoveCamera.INSTANCE, mPrefsMap.getBoolean("system_ui_lock_screen_hide_camera"));
-        initHook(RemoveSmartScreen.INSTANCE, mPrefsMap.getBoolean("system_ui_lock_screen_hide_smart_screen"));
         initHook(NoPassword.INSTANCE, mPrefsMap.getBoolean("system_ui_lock_screen_password_free"));
         initHook(LockScreenDoubleTapToSleep.INSTANCE, mPrefsMap.getBoolean("system_ui_lock_screen_double_lock"));
         initHook(HideLockscreenZenMode.INSTANCE, mPrefsMap.getBoolean("system_ui_lock_screen_not_disturb_mode"));
