+++ app/src/main/java/com/sevtinge/hyperceiler/module/hook/systemui/base/api/BaseReflectObject.kt
@@ -1,4 +1,4 @@
-package com.sevtinge.hyperceiler.module.hook.systemui.api
+package com.sevtinge.hyperceiler.module.hook.systemui.base.api
 
 abstract class BaseReflectObject(open val instance: Any) {
     override fun toString(): String = instance.toString()
