+++ app/src/main/java/com/sevtinge/hyperceiler/module/hook/systemui/base/api/MiuiStub.kt
@@ -1,4 +1,4 @@
-package com.sevtinge.hyperceiler.module.hook.systemui.api
+package com.sevtinge.hyperceiler.module.hook.systemui.base.api
 
 import android.content.Context
 import android.os.Handler
