+++ app/src/main/java/com/sevtinge/hyperceiler/module/hook/systemui/other/BrightnessPct.java
@@ -16,7 +16,7 @@
 
  * Copyright (C) 2023-2025 HyperCeiler Contributions
  */
-package com.sevtinge.hyperceiler.module.hook.systemui;
+package com.sevtinge.hyperceiler.module.hook.systemui.other;
 
 import static com.sevtinge.hyperceiler.module.base.tool.OtherTool.mPct;
 import static com.sevtinge.hyperceiler.module.base.tool.OtherTool.removePct;
