+++ app/src/main/java/com/sevtinge/hyperceiler/module/hook/systemui/other/DisableInfinitymodeGesture.kt
@@ -1,10 +1,10 @@
-package com.sevtinge.hyperceiler.module.hook.systemui
+package com.sevtinge.hyperceiler.module.hook.systemui.other
 
 import com.github.kyuubiran.ezxhelper.ClassUtils.loadClass
 import com.github.kyuubiran.ezxhelper.HookFactory.`-Static`.createHook
 import com.github.kyuubiran.ezxhelper.finders.MethodFinder.`-Static`.methodFinder
-import com.sevtinge.hyperceiler.module.base.*
-import com.sevtinge.hyperceiler.utils.devicesdk.*
+import com.sevtinge.hyperceiler.module.base.BaseHook
+import com.sevtinge.hyperceiler.utils.devicesdk.isMoreAndroidVersion
 
 
 object DisableInfinitymodeGesture : BaseHook() {
