+++ app/src/main/java/com/sevtinge/hyperceiler/module/hook/systemui/other/EnableVolumeBlur.kt
@@ -16,7 +16,7 @@
 
   * Copyright (C) 2023-2025 HyperCeiler Contributions
 */
-package com.sevtinge.hyperceiler.module.hook.systemui
+package com.sevtinge.hyperceiler.module.hook.systemui.other
 
 import com.sevtinge.hyperceiler.module.base.BaseHook
 import com.sevtinge.hyperceiler.utils.getValueByField
