+++ app/src/main/java/com/sevtinge/hyperceiler/module/hook/systemui/other/VolumeTimerValuesHook.java
@@ -16,7 +16,7 @@
 
   * Copyright (C) 2023-2025 HyperCeiler Contributions
 */
-package com.sevtinge.hyperceiler.module.hook.systemui;
+package com.sevtinge.hyperceiler.module.hook.systemui.other;
 
 import com.sevtinge.hyperceiler.R;
 import com.sevtinge.hyperceiler.XposedInit;
