+++ app/src/main/java/com/sevtinge/hyperceiler/ui/hooker/systemui/LockScreenSettings.java
@@ -26,9 +26,13 @@ import androidx.preference.SwitchPreference;
 import com.sevtinge.hyperceiler.R;
 import com.sevtinge.hyperceiler.ui.hooker.dashboard.DashboardFragment;
 
+import fan.preference.DropDownPreference;
+
 public class LockScreenSettings extends DashboardFragment {
+    SwitchPreference mHideLeftButton; // 隐藏左侧按钮
     SwitchPreference mPasswordFree; // 开机免输入密码
     SwitchPreference mBlockEditor; // 禁用长按进入锁屏编辑
+    DropDownPreference mHideLeftButtonNew; // 左侧按钮自定义
 
     @Override
     public int getPreferenceScreenResId() {
@@ -37,10 +41,16 @@ public class LockScreenSettings extends DashboardFragment {
 
     @Override
     public void initPrefs() {
+        final boolean moreAndroidVersion = isMoreAndroidVersion(35);
         mPasswordFree = findPreference("prefs_key_system_ui_lock_screen_password_free");
         mBlockEditor = findPreference("prefs_key_system_ui_lock_screen_block_editor");
 
-        mBlockEditor.setVisible(!isMoreAndroidVersion(35));
+        mHideLeftButton = findPreference("prefs_key_system_ui_lock_screen_hide_smart_screen");
+        mHideLeftButtonNew = findPreference("prefs_key_system_ui_lock_screen_bottom_left_button");
+
+        mBlockEditor.setVisible(!moreAndroidVersion);
+        mHideLeftButton.setVisible(!moreAndroidVersion);
+        mHideLeftButtonNew.setVisible(moreAndroidVersion);
 
         if (isDeviceEncrypted(requireContext())) {
             mPasswordFree.setChecked(false);
