+++ app/src/main/res/values/arrays.xml
@@ -594,7 +594,7 @@
 
     <string-array name="lock_screen_bottom_left_button">
         <item>@string/array_default</item>
-        <item>@string/hide</item>
+        <item>@string/array_hide</item>
         <item>@string/ic_qs_title_flashlight</item>
     </string-array>
