+++ app/src/main/res/values/strings.xml
@@ -330,7 +330,6 @@
     <string name="save_to_screen_title">Specification save location</string>
     <string name="only_new_version_can_do">Only newer versions supported</string>
     <string name="use_hyper_blur">Use Advanced textures</string>
-    <string name="hide">Hide</string>
     <string name="system_ui_statusbar_indicator_default_title">Default</string>
     <string name="system_ui_statusbar_indicator_expand_title">Expand</string>
     <string name="system_ui_statusbar_indicator_fixedcontent_width">Fix content width</string>
@@ -338,6 +337,7 @@
     <string name="system_ui_statusbar_indicator_bold">Bold</string>
     <!--Multi-list Array-->
     <string name="array_close">Close</string>
+    <string name="array_hide">Hide</string>
     <string name="array_default">Default</string>
     <string name="array_align_left">Left</string>
     <string name="array_align_right">Right</string>
